Cache server ACLs on a per-room basis.
